Chore: POST API에서 삭제와 수정 기능 분리
