+++ src/app/main/page.tsx
@@ -1,4 +1,5 @@
 "use client";
+
 import React from "react";
 import Link from "next/link";
 import Post from "@/components/post/PostItem";
